Finished Pause Menu + Minor Fixes
For the sake of simplicity and my sanity I've decided to remove the options button from the main menu and have the settings only be accessible from the pause menu. This resolves issue with not being able to go back to the pause or main menu depending on the previous screen.

Minor fixes include: adjusted background of main menu, Updated Death Screen with background.