Caves
Added a nested for loop to the GenerateMap() method that creates several long rectangles of air tiles which will be used as caves. The changeTile() method was modified so that no ores would spawn in the empty space of the cave and instead find another location. This is before knowing the current size of the enemy and player models and how they will behave once added. Further modifications will be made if necessary once branches merge.